backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
tensorflow-cpu==2.16.1
numpy==1.26.4
scikit-learn==1.5.2
pydantic==2.10.4
